A fix for block ID migration. We now also copy the old save's data section.
I found in Baby's First Space Race that when updating mods, block IDs would
migrate. Forge takes care of this when loading saves, but not when creating a
new world normally. This would lead to starting a new game with walls of
molten bronze and things like this. The apparent fix is to copy over the data
section from the save too. The server would notice discrepencies and correct
them as the game loads.